refactor: test result 기능 store로 이전, etc: test result json 수정
