+++ src/components/Test/Result/TestResultData.json
@@ -3,7 +3,7 @@
     "id": 0,
     "genre": ["드라마", "로맨스"],
     "name": "혜원(리틀포레스트)",
-    "img": "김태리1.jpg",
+    "img": "김태리.jpg",
     "title1": "따뜻한 감성을 가지고 있는 겸손한",
     "title2": "당신의 캐릭터는 김태리",
     "hashtag": ["활발한", "재밌는", "모범적인"],
@@ -13,7 +13,7 @@
     "id": 1,
     "genre": ["판타지"],
     "name": "해리 포터(해리 포터)",
-    "img": "해리포터1.jpeg",
+    "img": "해리포터.jpeg",
     "title1": "풍부한 상상력을 가지고 새로운것에 도전하는",
     "title2": "당신의 캐릭터는 해리 포터",
     "hashtag": ["활발한", "재밌는", "모범적인"],
@@ -23,7 +23,7 @@
     "id": 2,
     "genre": ["다큐멘터리"],
     "name": "캡틴 아메리카(캡틴 아메리카)",
-    "img": "캡틴아메리카1.jpg",
+    "img": "캡틴아메리카.jpg",
     "title1": "이상적인 세상을 만들어가는",
     "title2": "당신의 캐릭터는 캡틴 아메리카",
     "hashtag": ["활발한", "재밌는", "모범적인"],
@@ -33,7 +33,7 @@
     "id": 3,
     "genre": ["액션"],
     "name": "존윅(존윅)",
-    "img": "존윅1.jpeg",
+    "img": "존윅.jpeg",
     "title1": "사람과 관련된 뛰어난 통찰력을 가진",
     "title2": "당신의 캐릭터는 존 윅",
     "hashtag": ["활발한", "재밌는", "모범적인"],
@@ -43,7 +43,7 @@
     "id": 4,
     "genre": ["코미디"],
     "name": "형욱(럭키)",
-    "img": "형욱1.jpg",
+    "img": "형욱.jpg",
     "title1": "분위기를 고조시키는 우호적인",
     "title2": "당신의 캐릭터는 형욱",
     "hashtag": ["활발한", "재밌는", "모범적인"],
@@ -53,7 +53,7 @@
     "id": 5,
     "genre": ["애니메이션"],
     "name": "오오타 쿠니요시",
-    "img": "오타쿠1.jpg",
+    "img": "오타쿠.jpg",
     "title1": "한번 시작한 일은 끝까지 해내는",
     "title2": "당신의 캐릭터는 오오타 쿠니요시",
     "hashtag": ["활발한", "재밌는", "모범적인"],
@@ -63,7 +63,7 @@
     "id": 6,
     "genre": ["음악"],
     "name": "미아(라라랜드)",
-    "img": "미아1.jpg",
+    "img": "미아.jpg",
     "title1": "성실하고 온화하며 협조를 잘하는",
     "title2": "당신의 캐릭터는 미아",
     "hashtag": ["활발한", "재밌는", "모범적인"],
@@ -73,7 +73,7 @@
     "id": 7,
     "genre": ["호러"],
     "name": "애나벨",
-    "img": "애나벨1.jpg",
+    "img": "애나벨.jpg",
     "title1": "다양함을 선호하는",
     "title2": "당신의 캐릭터는 애나벨",
     "hashtag": ["활발한", "재밌는", "모범적인"],
@@ -83,7 +83,7 @@
     "id": 8,
     "genre": ["추리"],
     "name": "셜록 홈즈(셜록 홈즈)",
-    "img": "셜록1.jpg",
+    "img": "셜록.jpg",
     "title1": "논리적이고 뛰어난 상황적응력을 가진",
     "title2": "당신의 캐릭터는 셜록 홈즈",
     "hashtag": ["활발한", "재밌는", "모범적인"],
@@ -93,7 +93,7 @@
     "id": 9,
     "genre": ["로맨스"],
     "name": "메리(어바웃 타임)",
-    "img": "메리1.jpg",
+    "img": "메리.jpg",
     "title1": "따뜻한 감성을 가지고 있는 겸손한",
     "title2": "당신의 캐릭터는 메리",
     "hashtag": ["활발한", "재밌는", "모범적인"],
@@ -103,7 +103,7 @@
     "id": 10,
     "genre": ["SF"],
     "name": "네드(스파이더맨)",
-    "img": "네드1.jpg",
+    "img": "네드.jpg",
     "title1": "비평적인 관심을 가지고 있는 뛰어난",
     "title2": "당신의 캐릭터는 네드",
     "hashtag": ["활발한", "재밌는", "모범적인"],
@@ -113,7 +113,7 @@
     "id": 11,
     "genre": ["성인"],
     "name": "납득이(건축학개론)",
-    "img": "납득이1.png",
+    "img": "납득이.png",
     "title1": "열정적으로 새로운 관계를 만드는",
     "title2": "당신의 캐릭터는 납득이",
     "hashtag": ["활발한", "재밌는", "모범적인"],
